Fix: fixed refactor bug
